Refactor vertex layout logic into render context
Refactored VertexLayout logic from BgfxConsumer to a utility class such that it can be used by the render context.
