+++ Engine/Source/Runtime/Rendering/BgfxConsumer.cpp
@@ -1,5 +1,7 @@
 #include "BgfxConsumer.h"
 
+#include "Rendering/Utility/VertexLayoutUtility.h"
+
 namespace
 {
 
@@ -71,108 +73,6 @@ size_t GetSizeFromAttribType(const bgfx::AttribType::Enum attribType)
 		printf("Unknown attribute type!");
 		assert(false);	// Unknown type!
 		return 0;
-	}
-}
-
-void ConvertVertexLayout(const cd::VertexFormat::VertexAttributeLayout& vertexAttributeLayout, bgfx::VertexLayout& outVertexLayout)
-{
-	bgfx::Attrib::Enum vertexAttrib = bgfx::Attrib::Enum::Count;
-	bgfx::AttribType::Enum vertexAttribValue = bgfx::AttribType::Enum::Count;
-	bool normalized = false;
-
-	// Attribute Type
-	switch (vertexAttributeLayout.vertexAttributeType)
-	{
-	case cd::VertexAttributeType::Position :
-		vertexAttrib = bgfx::Attrib::Enum::Position;
-		break;
-	case cd::VertexAttributeType::Normal:
-		vertexAttrib = bgfx::Attrib::Enum::Normal;
-		normalized = true;
-		break;
-	case cd::VertexAttributeType::Tangent:
-		vertexAttrib = bgfx::Attrib::Enum::Tangent;
-		break;
-	case cd::VertexAttributeType::Bitangent:
-		vertexAttrib = bgfx::Attrib::Enum::Bitangent;
-		break;
-	case cd::VertexAttributeType::UV:
-		vertexAttrib = bgfx::Attrib::Enum::Count;
-		for (const bgfx::Attrib::Enum& textCoord : AllAttribUVTypes)
-		{
-			if (!outVertexLayout.has(textCoord))
-			{
-				vertexAttrib = textCoord;
-				break;
-			}
-		}
-		break;
-	case cd::VertexAttributeType::Color:
-		vertexAttrib = bgfx::Attrib::Enum::Count;
-		for (const bgfx::Attrib::Enum& color : AllAttribColorTypes)
-		{
-			if (!outVertexLayout.has(color))
-			{
-				vertexAttrib = color;
-				break;
-			}
-		}
-		break;
-	default:
-		vertexAttrib = bgfx::Attrib::Enum::Count;
-		break;
-	}
-
-	// Attribute Value
-	switch (vertexAttributeLayout.attributeValueType)
-	{
-	case cd::AttributeValueType::Uint8:
-		vertexAttribValue = bgfx::AttribType::Enum::Uint8;
-		break;
-	case cd::AttributeValueType::Float:
-		vertexAttribValue = bgfx::AttribType::Enum::Float;
-		break;
-	default:
-		vertexAttribValue = bgfx::AttribType::Enum::Count;
-		break;
-	}
-
-	assert(vertexAttrib != bgfx::Attrib::Enum::Count);
-	assert(vertexAttribValue != bgfx::AttribType::Enum::Count);
-	outVertexLayout.add(vertexAttrib, vertexAttributeLayout.attributeCount, vertexAttribValue, normalized);
-}
-
-std::string VertexAttributeTypeToString(const cd::VertexAttributeType& attribType)
-{
-	switch (attribType)
-	{
-		case cd::VertexAttributeType::Position:
-			return "Position";
-		case cd::VertexAttributeType::Normal:
-			return "Normal";
-		case cd::VertexAttributeType::Tangent:
-			return "Tangent";
-		case cd::VertexAttributeType::Bitangent:
-			return "Bitangent";
-		case cd::VertexAttributeType::UV:
-			return "UV";
-		case cd::VertexAttributeType::Color:
-			return "Color";
-		default:
-			return "Invalid Attribute Type!";
-	}
-}
-
-std::string AttributeValueTypeToString(const cd::AttributeValueType& attribValueType)
-{
-	switch (attribValueType)
-	{
-		case cd::AttributeValueType::Float:
-			return "Float";
-		case cd::AttributeValueType::Uint8:
-			return "Uint8";
-		default:
-			return "Invalid Attribute Value Type!";
 	}
 }
 
@@ -293,17 +193,8 @@ void BgfxConsumer::ConvertMeshesFromScene(const cd::SceneDatabase& sceneDatabase
 
 		// Convert vertex formats
  		bgfx::VertexLayout& vertexLayout = meshData.GetVertexLayout();
-		vertexLayout.begin();
-		for (const cd::VertexFormat::VertexAttributeLayout& vertexAttributeLayout : mesh.GetVertexFormat().GetVertexLayout())
-		{
-			printf("\t\tVA: (%s, %s, %d)\n",
-				VertexAttributeTypeToString(vertexAttributeLayout.vertexAttributeType).c_str(),
-				AttributeValueTypeToString(vertexAttributeLayout.attributeValueType).c_str(),
-				vertexAttributeLayout.attributeCount);
-			ConvertVertexLayout(vertexAttributeLayout, vertexLayout);
-		}
-		vertexLayout.end();
-
+		VertexLayoutUtility::CreateVertexLayout(vertexLayout, mesh.GetVertexFormat().GetVertexLayout(), true);
+		
 		// Convert vertices
 		std::vector<std::byte>& rawVertices = meshData.GetRawVertices(); 
 		rawVertices.resize(mesh.GetVertexCount() * vertexLayout.getStride());
